css
Change the colors to whatever you like.
Most of the buttons are implemented and working.
If you are having trouble it has to do with how extremely fast the chess player is. I have wondered about flaws in the game logic, yet it happens so fast I am not sure.
There may be a flaw in the chess player's openings that will let you take the king immediately, yet this also happens so fast it is hard to be sure.
Please enjoy and share!
And remember to tell your friends you have a chess engine!